move chrome api to src/chrome.js
